fix labels in example circuits
Instance variables and witness variables in the previous part always
individually start from index 1. This rule is broken in the first
fixed circuit.

In the second one this is also broken and in addition in the
simplified circuit we should rename the witnesses to go from 1 to 3
instead of having {1, 3, 4}.